Set Snowflake Id epoch to Kernel Genesis
Ahhh, this is one of those rare questions which must first be answered
with a poem https://dyeing.thebluebook.co.za/?stackedPages=%2Fjoy
this touches on an experience we had together around Devcon IV in
Prague. if memory serves, it was November 3rd, 2018 sometime in the late
afternoon. Let's call it 5:15pm in Prague.