Changed confirmation shown upon saving application
Added some additional context around what it means for the application to be "saved". I wanted to make the differences between save and submit clear by illustrating the action of leaving the website, coming back, and still needing to submit. Tried to maintain as close a character count as possible to the application submitted message to ensure it's not too long.